heppy batch working?
